refactor(url_builder): move URL building utilities to dedicated module
