+++ src/foo_opensubsonic.vcxproj
@@ -280,6 +280,7 @@
     <ClInclude Include="utils\subsonic_json_parser.h" />
     <ClInclude Include="utils\metadata_utils.h" />
     <ClInclude Include="utils\time_utils.h" />
+    <ClInclude Include="utils\url_builder.h" />
     <ClInclude Include="utils\utils.h" />
     <ClInclude Include="vfs.h" />
   </ItemGroup>
@@ -316,6 +317,9 @@
       <PrecompiledHeader>NotUsing</PrecompiledHeader>
     </ClCompile>
     <ClCompile Include="utils\time_utils.cpp">
+      <PrecompiledHeader>NotUsing</PrecompiledHeader>
+    </ClCompile>
+    <ClCompile Include="utils\url_builder.cpp">
       <PrecompiledHeader>NotUsing</PrecompiledHeader>
     </ClCompile>
     <ClCompile Include="utils\utils.cpp">
